requirements.txt
asgiref==3.11.1
attrs==26.1.0
Django==5.2.7
django-cors-headers==4.6.0
django-extensions==4.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
Faker>=25.0.0
pydotplus==2.0.2
pyparsing==3.3.2
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.2.0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2026.1
uritemplate==4.2.0
